# Performance Improvement - IsGenericTypeDefinedBy